mv gl plot interact tests to gl_plot_interact_test file:
- exclude that test file on circleci runs
- that way, gl plot interactions can be tested locally
